+++ app/celery/service_callback_tasks.py
@@ -5,6 +5,8 @@ from requests import HTTPError, RequestException, request
 
 from app import notify_celery, signing
 from app.config import QueueNames
+from app.dao.inbound_sms_dao import dao_get_inbound_sms_by_id
+from app.dao.service_inbound_api_dao import get_service_inbound_api_for_service
 from app.utils import DATETIME_FORMAT
 
 
@@ -53,6 +55,71 @@ def send_complaint_to_service(self, complaint_data):
         complaint["service_callback_api_bearer_token"],
         "send_complaint_to_service",
     )
+
+
+@notify_celery.task(bind=True, name="send-inbound-sms", max_retries=5, default_retry_delay=300)
+def send_inbound_sms_to_service(self, inbound_sms_id, service_id):
+    inbound_api = get_service_inbound_api_for_service(service_id=service_id)
+    if not inbound_api:
+        # No API data has been set for this service
+        return
+
+    inbound_sms = dao_get_inbound_sms_by_id(service_id=service_id, inbound_id=inbound_sms_id)
+    data = {
+        "id": str(inbound_sms.id),
+        # TODO: should we be validating and formatting the phone number here?
+        "source_number": inbound_sms.user_number,
+        "destination_number": inbound_sms.notify_number,
+        "message": inbound_sms.content,
+        "date_received": inbound_sms.provider_date.strftime(DATETIME_FORMAT),
+    }
+
+    try:
+        response = request(
+            method="POST",
+            url=inbound_api.url,
+            data=json.dumps(data),
+            headers={"Content-Type": "application/json", "Authorization": "Bearer {}".format(inbound_api.bearer_token)},
+            timeout=60,
+        )
+        current_app.logger.debug(
+            "send_inbound_sms_to_service sending %s to %s, response %s",
+            inbound_sms_id,
+            inbound_api.url,
+            response.status_code,
+        )
+        response.raise_for_status()
+    except RequestException as e:
+        current_app.logger.warning(
+            "send_inbound_sms_to_service failed for service_id: %s for inbound_sms_id: %s and url: %s. exception: %s",
+            service_id,
+            inbound_sms_id,
+            inbound_api.url,
+            e,
+        )
+        if not isinstance(e, HTTPError) or e.response.status_code >= 500:
+            try:
+                self.retry(queue=QueueNames.RETRY)
+            except self.MaxRetriesExceededError:
+                current_app.logger.error(
+                    (
+                        "Retry: send_inbound_sms_to_service has retried the max number of times "
+                        "for service: %s and inbound_sms %s"
+                    ),
+                    service_id,
+                    inbound_sms_id,
+                )
+        else:
+            current_app.logger.warning(
+                (
+                    "send_inbound_sms_to_service is not being retried for service_id: %s "
+                    "for inbound_sms id: %s and url: %s. exception: %s"
+                ),
+                service_id,
+                inbound_sms_id,
+                inbound_api.url,
+                e,
+            )
 
 
 def _send_data_to_service_callback_api(self, data, service_callback_url, token, function_name):
