+++ tests/app/celery/test_tasks.py
@@ -1,10 +1,8 @@
-import json
 import uuid
 from datetime import datetime, timedelta
 from unittest.mock import Mock, call
 
 import pytest
-import requests_mock
 from celery.exceptions import Retry
 from freezegun import freeze_time
 from notifications_utils.recipients import Row
@@ -13,7 +11,6 @@ from notifications_utils.template import (
     PlainTextEmailTemplate,
     SMSMessageTemplate,
 )
-from requests import RequestException
 from sqlalchemy.exc import SQLAlchemyError
 
 from app import signing
@@ -31,7 +28,6 @@ from app.celery.tasks import (
     save_email,
     save_letter,
     save_sms,
-    send_inbound_sms_to_service,
 )
 from app.config import QueueNames
 from app.constants import (
@@ -52,14 +48,12 @@ from app.v2.errors import TooManyRequestsError
 from tests.app import load_example_csv
 from tests.app.db import (
     create_api_key,
-    create_inbound_sms,
     create_job,
     create_letter_contact,
     create_notification,
     create_notification_history,
     create_reply_to_email,
     create_service,
-    create_service_inbound_api,
     create_service_with_defined_sms_sender,
     create_template,
     create_user,
@@ -1296,124 +1290,6 @@ def test_get_letter_template_instance(mocker, sample_job):
         "postcode",
         "address line 7",
     ]
-
-
-def test_send_inbound_sms_to_service_post_https_request_to_service(notify_api, sample_service):
-    inbound_api = create_service_inbound_api(
-        service=sample_service, url="https://some.service.gov.uk/", bearer_token="something_unique"
-    )
-    inbound_sms = create_inbound_sms(
-        service=sample_service,
-        notify_number="0751421",
-        user_number="447700900111",
-        provider_date=datetime(2017, 6, 20),
-        content="Here is some content",
-    )
-    data = {
-        "id": str(inbound_sms.id),
-        "source_number": inbound_sms.user_number,
-        "destination_number": inbound_sms.notify_number,
-        "message": inbound_sms.content,
-        "date_received": inbound_sms.provider_date.strftime(DATETIME_FORMAT),
-    }
-
-    with requests_mock.Mocker() as request_mock:
-        request_mock.post(inbound_api.url, json={}, status_code=200)
-        send_inbound_sms_to_service(inbound_sms.id, inbound_sms.service_id)
-    assert request_mock.call_count == 1
-    assert request_mock.request_history[0].url == inbound_api.url
-    assert request_mock.request_history[0].method == "POST"
-    assert request_mock.request_history[0].text == json.dumps(data)
-    assert request_mock.request_history[0].headers["Content-type"] == "application/json"
-    assert request_mock.request_history[0].headers["Authorization"] == "Bearer {}".format(inbound_api.bearer_token)
-
-
-def test_send_inbound_sms_to_service_does_not_send_request_when_inbound_sms_does_not_exist(notify_api, sample_service):
-    inbound_api = create_service_inbound_api(service=sample_service)
-    with requests_mock.Mocker() as request_mock:
-        request_mock.post(inbound_api.url, json={}, status_code=200)
-        with pytest.raises(SQLAlchemyError):
-            send_inbound_sms_to_service(inbound_sms_id=uuid.uuid4(), service_id=sample_service.id)
-
-    assert request_mock.call_count == 0
-
-
-def test_send_inbound_sms_to_service_does_not_sent_request_when_inbound_api_does_not_exist(
-    notify_api, sample_service, mocker
-):
-    inbound_sms = create_inbound_sms(
-        service=sample_service,
-        notify_number="0751421",
-        user_number="447700900111",
-        provider_date=datetime(2017, 6, 20),
-        content="Here is some content",
-    )
-    mocked = mocker.patch("requests.request")
-    send_inbound_sms_to_service(inbound_sms.id, inbound_sms.service_id)
-
-    assert mocked.call_count == 0
-
-
-def test_send_inbound_sms_to_service_retries_if_request_returns_500(notify_api, sample_service, mocker):
-    inbound_api = create_service_inbound_api(
-        service=sample_service, url="https://some.service.gov.uk/", bearer_token="something_unique"
-    )
-    inbound_sms = create_inbound_sms(
-        service=sample_service,
-        notify_number="0751421",
-        user_number="447700900111",
-        provider_date=datetime(2017, 6, 20),
-        content="Here is some content",
-    )
-
-    mocked = mocker.patch("app.celery.tasks.send_inbound_sms_to_service.retry")
-    with requests_mock.Mocker() as request_mock:
-        request_mock.post(inbound_api.url, json={}, status_code=500)
-        send_inbound_sms_to_service(inbound_sms.id, inbound_sms.service_id)
-
-    assert mocked.call_count == 1
-    assert mocked.call_args[1]["queue"] == "retry-tasks"
-
-
-def test_send_inbound_sms_to_service_retries_if_request_throws_unknown(notify_api, sample_service, mocker):
-    create_service_inbound_api(
-        service=sample_service, url="https://some.service.gov.uk/", bearer_token="something_unique"
-    )
-    inbound_sms = create_inbound_sms(
-        service=sample_service,
-        notify_number="0751421",
-        user_number="447700900111",
-        provider_date=datetime(2017, 6, 20),
-        content="Here is some content",
-    )
-
-    mocked = mocker.patch("app.celery.tasks.send_inbound_sms_to_service.retry")
-    mocker.patch("app.celery.tasks.request", side_effect=RequestException())
-
-    send_inbound_sms_to_service(inbound_sms.id, inbound_sms.service_id)
-
-    assert mocked.call_count == 1
-    assert mocked.call_args[1]["queue"] == "retry-tasks"
-
-
-def test_send_inbound_sms_to_service_does_not_retries_if_request_returns_404(notify_api, sample_service, mocker):
-    inbound_api = create_service_inbound_api(
-        service=sample_service, url="https://some.service.gov.uk/", bearer_token="something_unique"
-    )
-    inbound_sms = create_inbound_sms(
-        service=sample_service,
-        notify_number="0751421",
-        user_number="447700900111",
-        provider_date=datetime(2017, 6, 20),
-        content="Here is some content",
-    )
-
-    mocked = mocker.patch("app.celery.tasks.send_inbound_sms_to_service.retry")
-    with requests_mock.Mocker() as request_mock:
-        request_mock.post(inbound_api.url, json={}, status_code=404)
-        send_inbound_sms_to_service(inbound_sms.id, inbound_sms.service_id)
-
-    assert mocked.call_count == 0
 
 
 def test_process_incomplete_job_sms(mocker, sample_template):
