+++ tests/app/notifications/test_receive_notification.py
@@ -48,7 +48,9 @@ def mmg_post(client, data, auth=True, password="testkey"):
 
 
 def test_receive_notification_returns_received_to_mmg(client, mocker, sample_service_full_permissions):
-    mocked = mocker.patch("app.notifications.receive_notifications.tasks.send_inbound_sms_to_service.apply_async")
+    mocked = mocker.patch(
+        "app.notifications.receive_notifications.service_callback_tasks.send_inbound_sms_to_service.apply_async"
+    )
     prom_counter_labels_mock = mocker.patch("app.notifications.receive_notifications.INBOUND_SMS_COUNTER.labels")
     data = {
         "ID": "1234",
@@ -84,7 +86,9 @@ def test_receive_notification_returns_received_to_mmg(client, mocker, sample_ser
 def test_receive_notification_from_mmg_without_permissions_does_not_persist(
     client, mocker, notify_db_session, permissions
 ):
-    mocked = mocker.patch("app.notifications.receive_notifications.tasks.send_inbound_sms_to_service.apply_async")
+    mocked = mocker.patch(
+        "app.notifications.receive_notifications.service_callback_tasks.send_inbound_sms_to_service.apply_async"
+    )
     create_service_with_inbound_number(inbound_number="07111111111", service_permissions=permissions)
     data = {
         "ID": "1234",
@@ -116,7 +120,7 @@ def test_receive_notification_from_firetext_without_permissions_does_not_persist
     service = create_service_with_inbound_number(inbound_number="07111111111", service_permissions=permissions)
     mocker.patch("app.notifications.receive_notifications.dao_fetch_service_by_inbound_number", return_value=service)
     mocked_send_inbound_sms = mocker.patch(
-        "app.notifications.receive_notifications.tasks.send_inbound_sms_to_service.apply_async"
+        "app.notifications.receive_notifications.service_callback_tasks.send_inbound_sms_to_service.apply_async"
     )
     mocker.patch("app.notifications.receive_notifications.has_inbound_sms_permissions", return_value=False)
 
@@ -137,7 +141,7 @@ def test_receive_notification_without_permissions_does_not_create_inbound_even_w
     inbound_number = create_inbound_number("1", service_id=sample_service.id, active=True)
 
     mocked_send_inbound_sms = mocker.patch(
-        "app.notifications.receive_notifications.tasks.send_inbound_sms_to_service.apply_async"
+        "app.notifications.receive_notifications.service_callback_tasks.send_inbound_sms_to_service.apply_async"
     )
     mocked_has_permissions = mocker.patch(
         "app.notifications.receive_notifications.has_inbound_sms_permissions", return_value=False
@@ -315,7 +319,9 @@ def test_receive_notification_error_if_not_single_matching_service(client, notif
 
 
 def test_receive_notification_returns_received_to_firetext(notify_db_session, client, mocker):
-    mocked = mocker.patch("app.notifications.receive_notifications.tasks.send_inbound_sms_to_service.apply_async")
+    mocked = mocker.patch(
+        "app.notifications.receive_notifications.service_callback_tasks.send_inbound_sms_to_service.apply_async"
+    )
     prom_counter_labels_mock = mocker.patch("app.notifications.receive_notifications.INBOUND_SMS_COUNTER.labels")
 
     service = create_service_with_inbound_number(
@@ -338,7 +344,9 @@ def test_receive_notification_returns_received_to_firetext(notify_db_session, cl
 
 
 def test_receive_notification_from_firetext_persists_message(notify_db_session, client, mocker):
-    mocked = mocker.patch("app.notifications.receive_notifications.tasks.send_inbound_sms_to_service.apply_async")
+    mocked = mocker.patch(
+        "app.notifications.receive_notifications.service_callback_tasks.send_inbound_sms_to_service.apply_async"
+    )
     mocker.patch("app.notifications.receive_notifications.INBOUND_SMS_COUNTER")
 
     service = create_service_with_inbound_number(
@@ -364,7 +372,9 @@ def test_receive_notification_from_firetext_persists_message(notify_db_session, 
 
 
 def test_receive_notification_from_firetext_persists_message_with_normalized_phone(notify_db_session, client, mocker):
-    mocker.patch("app.notifications.receive_notifications.tasks.send_inbound_sms_to_service.apply_async")
+    mocker.patch(
+        "app.notifications.receive_notifications.service_callback_tasks.send_inbound_sms_to_service.apply_async"
+    )
     mocker.patch("app.notifications.receive_notifications.INBOUND_SMS_COUNTER")
 
     create_service_with_inbound_number(
@@ -385,7 +395,9 @@ def test_receive_notification_from_firetext_persists_message_with_normalized_pho
 
 
 def test_returns_ok_to_firetext_if_mismatched_sms_sender(notify_db_session, client, mocker):
-    mocked = mocker.patch("app.notifications.receive_notifications.tasks.send_inbound_sms_to_service.apply_async")
+    mocked = mocker.patch(
+        "app.notifications.receive_notifications.service_callback_tasks.send_inbound_sms_to_service.apply_async"
+    )
     mocker.patch("app.notifications.receive_notifications.INBOUND_SMS_COUNTER")
 
     create_service_with_inbound_number(
@@ -431,7 +443,9 @@ def test_strip_leading_country_code(number, expected):
     ],
 )
 def test_firetext_inbound_sms_auth(notify_db_session, notify_api, client, mocker, auth, keys, status_code):
-    mocker.patch("app.notifications.receive_notifications.tasks.send_inbound_sms_to_service.apply_async")
+    mocker.patch(
+        "app.notifications.receive_notifications.service_callback_tasks.send_inbound_sms_to_service.apply_async"
+    )
 
     create_service_with_inbound_number(
         service_name="b", inbound_number="07111111111", service_permissions=[EMAIL_TYPE, SMS_TYPE, INBOUND_SMS_TYPE]
@@ -458,7 +472,9 @@ def test_firetext_inbound_sms_auth(notify_db_session, notify_api, client, mocker
     ],
 )
 def test_mmg_inbound_sms_auth(notify_db_session, notify_api, client, mocker, auth, keys, status_code):
-    mocker.patch("app.notifications.receive_notifications.tasks.send_inbound_sms_to_service.apply_async")
+    mocker.patch(
+        "app.notifications.receive_notifications.service_callback_tasks.send_inbound_sms_to_service.apply_async"
+    )
 
     create_service_with_inbound_number(
         service_name="b", inbound_number="07111111111", service_permissions=[EMAIL_TYPE, SMS_TYPE, INBOUND_SMS_TYPE]
